+++ プロジェクト/Cliques_Engine/Project/Cliques_Engine/Cliques_Engine.vcxproj.filters
@@ -907,6 +907,9 @@
     <Filter Include="ヘッダー ファイル\レンダリング\定義\ブレンド">
       <UniqueIdentifier>{385fffde-b828-4ee6-9c67-4e709759ae50}</UniqueIdentifier>
     </Filter>
+    <Filter Include="ヘッダー ファイル\レンダリング\定義\深度ステンシル">
+      <UniqueIdentifier>{193868c4-acfd-477a-a45b-47ca8393b04c}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="C_Windows_OS_Management_System.h">
@@ -1073,9 +1076,6 @@
     </ClInclude>
     <ClInclude Include="C_Texture_Manager.h">
       <Filter>ヘッダー ファイル\アセット\テクスチャ\制御</Filter>
-    </ClInclude>
-    <ClInclude Include="Depth_Stencil_Setting_Inform.h">
-      <Filter>ヘッダー ファイル\レンダリング\定義</Filter>
     </ClInclude>
     <ClInclude Include="Rasterizer_Setting_Inform.h">
       <Filter>ヘッダー ファイル\レンダリング\定義</Filter>
@@ -1296,14 +1296,8 @@
     <ClInclude Include="S_Unique_Buffer_Slot.h">
       <Filter>ヘッダー ファイル\アセット\シェーダー\定義</Filter>
     </ClInclude>
-    <ClInclude Include="E_DEPTH_STENCIL_BUFFER_PIXEL_SIZE.h">
-      <Filter>ヘッダー ファイル\レンダリング\定義</Filter>
-    </ClInclude>
     <ClInclude Include="S_Create_Depth_Stencil_Buffer_Inform.h">
       <Filter>ヘッダー ファイル\レンダリング\API\生成用データ定義</Filter>
-    </ClInclude>
-    <ClInclude Include="E_DEPTH_STENCIL_BUFFER_PIXEL_BYTE_FORMAT.h">
-      <Filter>ヘッダー ファイル\レンダリング\定義</Filter>
     </ClInclude>
     <ClInclude Include="C_Animation_Model_Loader.h">
       <Filter>ヘッダー ファイル\アセット\3Dモデル\アニメーションモデル\ロードシステム</Filter>
@@ -1508,6 +1502,24 @@
     </ClInclude>
     <ClInclude Include="E_Rendering_Draw_Color.h">
       <Filter>ヘッダー ファイル\レンダリング\定義\ブレンド</Filter>
+    </ClInclude>
+    <ClInclude Include="Depth_Stencil_Setting_Inform.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
+    </ClInclude>
+    <ClInclude Include="E_DEPTH_STENCIL_BUFFER_PIXEL_BYTE_FORMAT.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
+    </ClInclude>
+    <ClInclude Include="E_DEPTH_STENCIL_BUFFER_PIXEL_SIZE.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
+    </ClInclude>
+    <ClInclude Include="E_Depth_Mode.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
+    </ClInclude>
+    <ClInclude Include="E_Depth_Write_Rule.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
+    </ClInclude>
+    <ClInclude Include="E_Stencil_Activate_Setting.h">
+      <Filter>ヘッダー ファイル\レンダリング\定義\深度ステンシル</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
