+++ プロジェクト/Cliques_Engine/Project/Cliques_Engine/Depth_Stencil_Setting_Inform.h
@@ -1,5 +1,5 @@
 //☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆//
-// 詳細   ：深度ステンシルの定義を行うファイル
+// 詳細   ：深度ステンシル用の定義を一括で読み込むためのファイル
 // 説明   ：
 // 作成者 ：髙坂龍誠
 //☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆☆//
@@ -10,42 +10,11 @@
 #define D_INCLUDE_GUARD_DEPTH_STENCIL_INFORM_H_FILE
 
 
-// ☆ ネームスペース ☆ //
+// ☆ ファイルひらき ☆ //
+#include "E_Depth_Mode.h"
+#include "E_Depth_Write_Rule.h"
+#include "E_Stencil_Activate_Setting.h"
 
-// 深度ステンシルの定義用の情報を呼び出すための名前
-namespace RENDERING::INFORM::DEPTH_STENCIL
-{
-	// ☆ 列挙 ☆ //
-
-	// 深度の書き込み方法の列挙
-	enum class E_DEPTH_MODE : unsigned char
-	{
-		e_NO,	// 深度は無効
-		e_ZERO,	// 深度は有効、元の深度値が０の部分にのみ書き込む
-		e_ONE,	// 深度は有効、全ての範囲に書き込む
-	};
-
-	// 深度の書き込みルールの列挙
-	enum class E_DEPTH_WRITE_RULE : unsigned char
-	{
-		e_DONT,			// 書き込まない
-		e_ALWAYS,		// 常に書き込む
-		e_EQUAL,		// 比較元と同じ値の時のみ書き込む
-		e_NOT_EQUAL,	// 比較元と違う値の時のみ書き込む
-		e_GREATER,		// 比較元の値を超える時のみ書き込む
-		e_LESS,			// 比較元の値未満の時のみ書き込む
-
-		e_LESS_EQUAL,		// 比較元の値以下の時のみ書き込む
-		e_GREATERE_EQUAL,	// 比較元の値以上の時のみ書き込む
-	};
-
-	// ステンシルが有効か無効を示す列挙
-	enum class E_STENCIL_IS : bool
-	{
-		e_NO_ACTIVE = false,	// 無効
-		e_ACTIVE = true,		// 有効
-	};
-}
 
 #endif // !D_INCLUDE_GUARD_DEPTH_STENCIL_INFORM_H_FILE
