+++ プロジェクト/Cliques_Engine/Project/Cliques_Engine/S_Create_Rendering_Graphics_Setting_Inform.h
@@ -48,7 +48,7 @@ namespace RENDERING::API::CREATE
 
 		RENDERING::INFORM::DEPTH_STENCIL::E_DEPTH_WRITE_RULE write_rule = RENDERING::INFORM::DEPTH_STENCIL::E_DEPTH_WRITE_RULE::e_ALWAYS;	// 深度値書き込みルール
 
-		RENDERING::INFORM::DEPTH_STENCIL::E_STENCIL_IS stencil_activate = RENDERING::INFORM::DEPTH_STENCIL::E_STENCIL_IS::e_NO_ACTIVE;	// ステンシルを有効にするかどうか
+		RENDERING::INFORM::DEPTH_STENCIL::E_STENCIL_ACTIVATE_SETTING stencil_activate = RENDERING::INFORM::DEPTH_STENCIL::E_STENCIL_ACTIVATE_SETTING::e_NO_ACTIVE;	// ステンシルを有効にするかどうか
 	};
 
 	// ラスタライザ設定を生成するためのデータの構造体
